通信エラーダイアログを閉じるを別ファイルに切り出した
